requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Config
python-dotenv==1.0.1

# Multi-agent orchestration
langgraph>=0.2.0
langchain-core>=0.2.0

# Observability / tracing
langsmith>=0.1.0

# Async SQLite (evolution tracker)
aiosqlite>=0.20.0

# Frontend
streamlit==1.41.1
plotly==5.24.1
pandas>=2.0.0
requests==2.32.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Linting / type checking
ruff==0.8.4
mypy==1.13.0
